requirements.txt
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
python-decouple==3.6
psycopg2
django-picklefield
requests==2.27.1